feat(grpo): add agentic tool rollouts and persistent GLM decode
Support tool calls across native-colocate model families using checkpoint
protocols and fallback templates. Preserve generated tokens, policy logprobs
and tool-result masks through Verifiers multi-turn rollouts, including native
Qwen serving and streaming responses.

Add persistent GLM decode state, registered sparse DSA kernels and recurrent
FLA KDA execution. Align rollout and training numerics for long GRPO sequences,
and reset cached state when training changes the policy.

Include a local ToolEnv example, an authentic GLM dummy tokenizer/template,
parser and training regressions, and updated capability documentation.

Validation: CPU regressions and C++ parser/schema tests passed. GPU tests used
only GPU 7, including real Qwen3.5 tool use and streaming, a complete native
GRPO CLI run, and a tiny GLM tool trajectory with a nonzero training update.

+++ surogate/kernels/triton/fla_kda/wy.py
@@ -75,7 +75,8 @@ def recompute_w_u_fwd_kda_kernel(
         m_v = m_t[:, None] & (o_v[None, :] < V)
         p_v = v + o_t[:, None] * (HV*V) + o_v[None, :]
         p_u = u + o_t[:, None] * (HV*V) + o_v[None, :]
-        b_v = tl.load(p_v, mask=m_v, other=0.0)
+        # Surogate: preserve FP32 WY intermediates with BF16 model values.
+        b_v = tl.load(p_v, mask=m_v, other=0.0).to(b_A.dtype)
         b_vb = (b_v * b_b[:, None]).to(b_v.dtype)
         b_u = tl.dot(b_A, b_vb)
         tl.store(p_u, b_u.to(p_u.dtype.element_ty), mask=m_v)
